+++ flake8_trio.py
@@ -14,7 +14,7 @@ import tokenize
 from typing import Any, Generator, Iterable, List, Optional, Tuple, Type, Union
 
 # CalVer: YY.month.patch, e.g. first release of July 2022 == "22.7.1"
-__version__ = "22.7.4"
+__version__ = "22.7.5"
 
 
 Error = Tuple[int, int, str, Type[Any]]